Cast negated values to "uint" before casting up to "ulong", to prevent sign-extension.

+++ runtime/System/Int16.cs
@@ -79,7 +79,8 @@ public struct Int16 : IComparable, IFormattable
 					else
 					{
 						return NumberFormatter.FormatFixedPoint
-									((ulong)(-value__), 0, 0, true, format,
+									((ulong)(uint)(-value__), 0, 0,
+									 true, format,
 									 NumberFormatInfo.GetInstance(provider));
 					}
 				}
